% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Histo_unique}
\alias{Histo_unique}
\title{Histogram function
Histograms of unique transcripts and genes}
\usage{
Histo_unique()
}
\description{
Histogram function
Histograms of unique transcripts and genes
}
