% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ST.QC}
\alias{ST.QC}
\title{ST-package
conversion of tsv files and alignment matrix to data frame with filtered data and HTGC nomenclature}
\usage{
ST.QC()
}
\description{
ST-package
conversion of tsv files and alignment matrix to data frame with filtered data and HTGC nomenclature
}
