% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subset_data}
\alias{subset_data}
\title{subset the expression datasetsusing the spots under tissue
removes ambiguous, non-expressed genes and prints general stats}
\usage{
subset_data(data, alignment)
}
\description{
subset the expression datasetsusing the spots under tissue
removes ambiguous, non-expressed genes and prints general stats
}
